Se refactoriza el codigo: eliminando segmentos
