--- src/app/dashboard/gobiernoDatos/listDashboard.json
@@ -1,36 +1,0 @@
-{
-  "ListDashboard": [
-    {
-      "cUrl": "https://lookerstudio.google.com/embed/reporting/6e29a280-d5b7-4430-8c0a-65a932833543/page/D6jeD",
-      "cNombreDashboard": "Contratos",
-      "cDescripcion": "Dashboard para contratos",
-      "lEstado": true,
-      "idDashboard": 1,
-      "lstPerfiles": [
-        {"cPerfil":"Gerente"},
-        {"cPerfil":"Seguridad TI"}
-      ]
-    },
-    {
-      "cUrl": "https://lookerstudio.google.com/embed/reporting/8f152ca5-a1fb-4d29-869d-ce54ecef2e50/page/0fIVD",
-      "cNombreDashboard": "Consolidado iniciativas PI",
-      "cDescripcion": "Dashboard para contratos",
-      "lEstado": true,
-      "idDashboard": 2,
-      "lstPerfiles": [
-        {"cPerfil":"Gerente"}
-      ]
-    },
-    {
-      "cUrl": "https://lookerstudio.google.com/embed/reporting/797196b5-1cba-4e10-9c18-201bb90df336/page/p_iudtpg9pdd",
-      "cNombreDashboard": "Auditoría",
-      "cDescripcion": "Dashboard para contratos",
-      "lEstado": false,
-      "idDashboard": 3,
-      "lstPerfiles": [
-        {"cPerfil":"Sub-Gerentes"},
-        {"cPerfil":"Seguridad TI"}
-      ]
-    }
-  ]
-}
